+++ src/core/DfT.ZEV.Core.Application/Processes/ProcessService.cs
@@ -2,6 +2,7 @@ using DfT.ZEV.Core.Domain.Processes.Models;
 using DfT.ZEV.Core.Domain.Processes.Services;
 using DfT.ZEV.Core.Domain.Processes.Values;
 using DfT.ZEV.Core.Infrastructure.Repositories;
+using DfT.ZEV.Core.Infrastructure.UnitOfWork;
 
 namespace DfT.ZEV.Core.Application.Processes;
