+++ src/core/DfT.ZEV.Core.Infrastructure/DfT.ZEV.Core.Infrastructure.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\common\DfT.ZEV.Common\DfT.ZEV.Common.csproj" />
         <ProjectReference Include="..\DfT.ZEV.Core.Domain\DfT.ZEV.Core.Domain.csproj"/>
     </ItemGroup>
