+++ src/core/DfT.ZEV.Core.Infrastructure/UnitOfWork/UnitOfWork.cs
@@ -3,11 +3,12 @@ using Microsoft.EntityFrameworkCore.Storage;
 using DfT.ZEV.Core.Domain.Processes.Services;
 using DfT.ZEV.Core.Domain.Vehicles.Services;
 using DfT.ZEV.Core.Infrastructure.Persistence;
+using DfT.ZEV.Core.Infrastructure.Repositories;
 
-namespace DfT.ZEV.Core.Infrastructure.Repositories;
+namespace DfT.ZEV.Core.Infrastructure.UnitOfWork;
 
-/// <inheritdoc cref="DfT.ZEV.Core.Infrastructure.Repositories.IUnitOfWork" />
-public class UnitOfWork : IUnitOfWork, IDisposable
+/// <inheritdoc cref="IUnitOfWork" />
+internal sealed class UnitOfWork : IUnitOfWork, IDisposable
 {
     private readonly AppDbContext _context;
     private bool _disposed;
