+++ src/core/tests/DfT.ZEV.Core.Application.Tests/Processes/ProcessServiceTests.cs
@@ -4,6 +4,7 @@ using DfT.ZEV.Core.Application.Processes;
 using DfT.ZEV.Core.Domain.Processes.Models;
 using DfT.ZEV.Core.Domain.Processes.Values;
 using DfT.ZEV.Core.Infrastructure.Repositories;
+using DfT.ZEV.Core.Infrastructure.UnitOfWork;
 using Moq;
 
 namespace DfT.ZEV.Core.Application.Tests.Processes;
