+++ src/services/apis/scheme-data/DfT.ZEV.Services.SchemeData.Api/DfT.ZEV.Services.SchemeData.Api.csproj
@@ -15,4 +15,10 @@
       <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Reference Include="DfT.ZEV.Core.Infrastructure">
+        <HintPath>..\..\..\..\core\tests\DfT.ZEV.Core.Infrastructure.Tests\bin\Debug\net6.0\DfT.ZEV.Core.Infrastructure.dll</HintPath>
+      </Reference>
+    </ItemGroup>
+
 </Project>
