+++ src/services/apis/scheme-data/DfT.ZEV.Services.SchemeData.Api/Features/Processes/DeleteProcess/DeleteProcessHandler.cs
@@ -1,11 +1,11 @@
 using Microsoft.AspNetCore.Mvc;
-using DfT.ZEV.Core.Infrastructure.Repositories;
+using DfT.ZEV.Core.Infrastructure.UnitOfWork;
 
 namespace DfT.ZEV.Services.SchemeData.Api.Features.Processes.DeleteProcess;
 
 public class DeleteProcessHandler
 {
-    public static async Task<IResult> HandleAsync([FromServices] UnitOfWork unitOfWork, [FromRoute] Guid processId,
+    public static async Task<IResult> HandleAsync([FromServices] IUnitOfWork unitOfWork, [FromRoute] Guid processId,
         CancellationToken cancellationToken = default)
     {
         var process = await unitOfWork.Processes.GetByIdAsync(processId, cancellationToken);
