+++ src/services/apis/scheme-data/DfT.ZEV.Services.SchemeData.Api/Features/Processes/GetProcessById/GetProcessByIdHandler.cs
@@ -1,5 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
-using DfT.ZEV.Core.Infrastructure.Repositories;
+using DfT.ZEV.Core.Infrastructure.UnitOfWork;
 
 namespace DfT.ZEV.Services.SchemeData.Api.Features.Processes.GetProcessById;
