+++ src/services/apis/scheme-data/DfT.ZEV.Services.SchemeData.Api/Setup.cs
@@ -1,11 +1,9 @@
 using System.Text.Json.Serialization;
 using DfT.ZEV.Common.Configuration;
-using DfT.ZEV.Core.Infrastructure.Persistence;
-using DfT.ZEV.Core.Infrastructure.Repositories;
+using DfT.ZEV.Core.Infrastructure;
 using Microsoft.AspNetCore.Http.Json;
 
 using DfT.ZEV.Services.SchemeData.Api.Features.Processes;
-using Microsoft.EntityFrameworkCore;
 
 namespace DfT.ZEV.Services.SchemeData.Api;
 
@@ -19,12 +17,8 @@ public static class Setup
         builder.Services.Configure<JsonOptions>(o => o.SerializerOptions.Converters.Add(new JsonStringEnumConverter()));
 
         var postgresSettings = builder.Services.ConfigurePostgresSettings(builder.Configuration);
-
-        builder.Services.AddDbContext<AppDbContext>(opt =>
-        {
-            opt.UseNpgsql(postgresSettings.ConnectionString,
-                conf => { conf.EnableRetryOnFailure(5, TimeSpan.FromSeconds(20), new List<string> { "4060" }); });
-        });
+        
+        builder.Services.AddDbContext(postgresSettings);
         
         //to-do: add serilog from commons
         //builder.Services.AddSerilog(builder.Configuration);
